solved use case 6

+++ StudentRegistrationSystem/src/main/java/registrationsystem/service/impl/AcademicBlockServiceImpl.java
@@ -5,6 +5,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import registrationsystem.domain.AcademicBlock;
+import registrationsystem.exception.CourseExceptionHandler;
 import registrationsystem.repository.AcademicBlockRepository;
 import registrationsystem.service.AcademicBlockService;
 import registrationsystem.service.dto.AcademicBlockDTO;
@@ -22,19 +23,24 @@ public class AcademicBlockServiceImpl implements AcademicBlockService {
 
     @Override
     public void addBlock(AcademicBlock block) {
+        log.info("Saving Block with Id :" + block.toString());
         repository.save(block);
     }
 
     @Override
     public AcademicBlockDTO getBlock(Long id) {
         var block = repository.findById(id).get();
+        if(block == null){
+            log.info("Block with Id :" + id + " not found");
+            throw new CourseExceptionHandler("Block with Id :" + id+ " not found");
+        }
         return modelMapper.map(block, AcademicBlockDTO.class);
     }
 
     @Override
     public Collection<AcademicBlockDTO> getAllBlocks() {
         var allBlocks = repository.findAll();
-
+        log.info("fetching all blocks :");
         return allBlocks.stream()
                 .map(block -> modelMapper.map(block, AcademicBlockDTO.class))
                 .collect(Collectors.toList());
@@ -52,13 +58,21 @@ public class AcademicBlockServiceImpl implements AcademicBlockService {
             updateBlock.setStartDate(block.getStartDate());
             updateBlock.setEndDate(block.getEndDate());
             updateBlock.setCourseOfferings(block.getCourseOfferings());
+            log.info("updating Block with Id :" + id );
             repository.save(updateBlock);
+        } else {
+            throw new CourseExceptionHandler("Block with Id :" + id + " not found");
         }
         return modelMapper.map(updateBlock, AcademicBlockDTO.class);
     }
 
     @Override
     public void deleteBlock(Long id) {
+        var toDelete = repository.findById(id);
+        if(toDelete == null){
+            log.info("Deleting Block with Id :");
+            throw new CourseExceptionHandler("Block with Id :" + id +" not found");
+        }
         repository.deleteById(id);
     }
 }
